custom Hook으로로직 분리
